requirements.txt
streamlit
folium
networkx
matplotlib.pyplot
matplotlib.cm
seaborn
numpy
pandas